cree el end poin de cart que crea y borra carritos con id, un objeto products y quantity

+++ app.js
@@ -37,10 +37,14 @@ app.get('/productos/:id', (req, res) => {
 });
 
 app.delete ('/productos/delete/:id', (req, res)=>{
-
+   if(productManager.getProductById){
     productManager.deleteProduct(parseInt(req.params.id));
-    return res.status(404).json({ mensaje: 'Producto borrado' });
-})
+    return res.status(200).json({ mensaje: 'Producto borrado' });
+   }
+   else {
+    return res.status(404).json({ mensaje: 'Producto no encontrado' });
+}
+});
 
 
 // manejo de carts---------------------
@@ -48,22 +52,25 @@ app.delete ('/productos/delete/:id', (req, res)=>{
 app.post('/cart', (req, res) => {
 
     cartManager.addCart();
-    res.status(200).json({ success: 'El carrito fue creado' });
+    res.status(200).json({ success: `carrito ${cartManager.nextId-1} creado` });
 });
 
 app.delete ('/cart/delete/:id', (req, res)=>{
     const cartId = parseInt(req.params.id);
-    const deleted = cartManager.deleteCart(cartId);
+    const cartExists = cartManager.cartExists (cartId)
 
-    if (deleted) {
-        return res.status(200).json({ mensaje: 'Carrito borrado exitosamente' });
+    if (cartExists) {
+        cartManager.deleteCart(cartId);
+        return res.status(200).json({ mensaje: `Carrito ${cartId} fue borrado` });
     } else {
-        return res.status(404).json({ mensaje: 'Carrito no encontrado' });
+        return res.status(404).json({ mensaje: `Carrito ${cartId} no encontrado`});
+        
     }
-    // cartManager.deleteCart(parseInt(req.params.id));
-    // return res.status(404).json({ mensaje: 'carrito borrado' });
 });
 
+
+  // cartManager.deleteCart(parseInt(req.params.id));
+    // return res.status(404).json({ mensaje: 'carrito borrado' });
